# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sharon Online API
  version: 1.0.0
  description: 25-minute 1-on-1 ESL marketplace API.
paths:
...
  /api/v1/bookings/{booking_id}/review/:
    post:
      operationId: v1_bookings_review_create
      description: Allows a student to submit a 1-5 star rating and optional written
        review.
      parameters:
      - in: path
        name: booking_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/bookings/{id}/:
    get:
      operationId: v1_bookings_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetail'
          description: ''
...
  /api/v1/student/bookings/{id}/review/:
    post:
      operationId: v1_student_bookings_review_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
    AccentEnum:
      enum:
      - ZA
      - UK
      - US
      - OTHER
      type: string
      description: |-
        * `ZA` - South African
        * `UK` - British
        * `US` - American
        * `OTHER` - International
...
    BookingDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        booking_reference:
          type: string
          readOnly: true
        teacher:
          allOf:
          - $ref: '#/components/schemas/TeacherList'
          readOnly: true
        student:
          allOf:
          - $ref: '#/components/schemas/BookingStudent'
          readOnly: true
        material:
          type: string
          format: uuid
          nullable: true
        material_slug:
          type: string
          nullable: true
          readOnly: true
        material_title:
          type: string
          nullable: true
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          default: pending_payment
        start_time_utc:
          type: string
          format: date-time
        end_time_utc:
          type: string
          format: date-time
        local_date:
          type: string
          readOnly: true
        local_start_time:
          type: string
          readOnly: true
        local_end_time:
          type: string
          readOnly: true
        viewer_timezone:
          type: string
          readOnly: true
        price_usd:
          type: number
          format: double
          readOnly: true
        price_zar:
          type: number
          format: double
          readOnly: true
        lock_expires_at:
          type: string
          nullable: true
          readOnly: true
        zoom_url:
          type: string
          readOnly: true
        zoom_join_url:
          type: string
          readOnly: true
        zoom_start_url:
          type: string
          readOnly: true
        zoom_meeting_id:
          type: string
          maxLength: 64
        zoom_password:
          type: string
          maxLength: 32
        student_rating:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
          description: 1 to 5 star rating
        student_review:
          type: string
          description: Written review visible to admin and teacher
        memo:
          allOf:
          - $ref: '#/components/schemas/LessonMemo'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - booking_reference
      - created_at
      - end_time_utc
      - id
      - local_date
      - local_end_time
      - local_start_time
      - lock_expires_at
      - material_slug
      - material_title
      - memo
      - price_usd
      - price_zar
      - start_time_utc
      - student
      - teacher
      - viewer_timezone
      - zoom_join_url
      - zoom_start_url
      - zoom_url
    BookingStudent:
      type: object
      description: Minimal student view embedded in a booking. Contact details are
        only for the student themselves/staff.
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          readOnly: true
        first_name:
          type: string
        timezone:
          type: string
        country:
          type: string
      required:
      - country
      - first_name
      - full_name
      - id
      - timezone
...
    LessonMemo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        booking:
          type: string
          format: uuid
          readOnly: true
        teacher:
          type: string
          format: uuid
          readOnly: true
        student:
          type: string
          format: uuid
          readOnly: true
        feedback_text:
          type: string
          description: Grammar corrections, speaking feedback, overall commentary
        vocabulary_words:
          description: 'List of words: [{''word'': ''resilience'', ''definition'':
            ''...''}]'
        pronunciation_notes:
          type: string
        homework:
          type: string
        submitted_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - booking
      - feedback_text
      - id
      - student
      - submitted_at
      - teacher
...
    StatusEnum:
      enum:
      - pending_payment
      - confirmed
      - in_progress
      - completed
      - cancelled
      - disputed
      - interrupted_power
      - student_no_show
      - teacher_no_show
      - completed_pending_memo
      - completed_memo_forfeited
      type: string
      description: |-
        * `pending_payment` - Pending Payment
        * `confirmed` - Confirmed
        * `in_progress` - In Progress
        * `completed` - Completed
        * `cancelled` - Cancelled
        * `disputed` - Disputed
        * `interrupted_power` - Interrupted (Power Outage)
        * `student_no_show` - Student No Show
        * `teacher_no_show` - Teacher No Show
        * `completed_pending_memo` - Completed (Pending Memo)
        * `completed_memo_forfeited` - Completed (Memo Forfeited)
...
    TeacherList:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        user_id:
          type: string
          format: uuid
          readOnly: true
        full_name:
          type: string
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        headline:
          type: string
          description: e.g. Certified TEFL Tutor · 5+ Yrs Experience
          maxLength: 255
        accent:
          $ref: '#/components/schemas/AccentEnum'
        intro_video_thumbnail:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        avatar_url:
          type: string
          readOnly: true
        intro_audio_url:
          type: string
          readOnly: true
        rating_avg:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
        rating_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        price_per_25min_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,4}(?:\.\d{0,2})?$
        specialties:
          description: 'List of tags: [''FreeTalk'', ''Business English'', ''Daily
            News'', ''TOEIC'']'
        country:
          type: string
          readOnly: true
        is_verified:
          type: boolean
      required:
      - avatar_url
      - country
      - first_name
      - full_name
      - id
      - intro_audio_url
      - last_name
      - user_id
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT